Remove the HostIO-caching feature
Removing this feature reduces the contract binary size and the gas usage
when the cache is not hit. The caching added an overhead that required
it to hit dozens of times to be worth it.

Users are advised to add caching in the form of local variables if the
contract calls the same HostIO multiple times.

Close STY-235